# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BetTopia gambling platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases/{id}:
    get:
      operationId: getCaseById
      tags: [cases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      summary: Get a case by ID
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{id}/open:
    post:
      operationId: openCase
      tags: [cases]
      summary: Open a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case opened, item won
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseOpenResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get recent chat messages
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CaseItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        imageUrl:
          type: string
        rarity:
          type: string
        value:
          type: number
        chance:
          type: number
        color:
          type: string
      required: [id, name, imageUrl, rarity, value, chance, color]
    Case:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        category:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CaseItem"
      required: [id, name, imageUrl, price, category, items]
    CaseOpenResult:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/CaseItem"
        newBalance:
          type: number
      required: [item, newBalance]
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        avatar:
          type: string
        message:
          type: string
        createdAt:
          type: string
      required: [id, userId, username, message, createdAt]
    SendMessageBody:
      type: object
      properties:
        message:
          type: string
      required: [message]
...